Brings database queues to a runnable state
