+++ Sources/Alchemy/Jobs/Helpers/Frequency.swift
@@ -1,15 +1,16 @@
 import Foundation
 
-/// Represents a frequency that occurs at a `rate` and may have specific requirements for when it
-/// should start running, such as "every day at 9:30 am".
+/// Represents a frequency that occurs at a `rate` and may have
+/// specific requirements for when it should start running,
+/// such as "every day at 9:30 am".
 public protocol Frequency {
     /// The interval at which this `Frequency` is repeated.
     var rate: TimeAmount { get }
-    
+
     /// An interval until the next time this frequency can begin.
     ///
-    /// - Returns: a `TimeAmount` representing the interval between now and when this frequency's
-    ///            start time will next occur.
+    /// - Returns: A `TimeAmount` representing the interval between
+    ///   now and when this frequency's start time will next occur.
     func timeUntilNext() -> TimeAmount
 }
 
@@ -19,43 +20,34 @@ public protocol Frequency {
 public protocol TimeUnit {
     /// Returns a `TimeAmount` for a measure of time with this unit.
     ///
-    /// - Parameter value: the measurement of time.
-    /// - Returns an interval representing `value` units of this `TimeUnit`.
+    /// - Parameter value: The measurement of time.
+    /// - Returns An interval representing `value` units of this
+    ///   `TimeUnit`.
     static func time(_ value: Int64) -> TimeAmount
 }
 
 /// A week of time.
 public struct WeekUnit: TimeUnit {
-    // MARK: TimeUnit
-    
     public static func time(_ value: Int64) -> TimeAmount { .hours(value * 24 * 7) }
 }
 
 /// A day of time.
 public struct DayUnit: TimeUnit {
-    // MARK: TimeUnit
-    
     public static func time(_ value: Int64) -> TimeAmount { .hours(value * 24) }
 }
 
 /// An hour of time.
 public struct HourUnit: TimeUnit {
-    // MARK: TimeUnit
-    
     public static func time(_ value: Int64) -> TimeAmount { .hours(value) }
 }
 
 /// A minute of time.
 public struct MinuteUnit: TimeUnit {
-    // MARK: TimeUnit
-    
     public static func time(_ value: Int64) -> TimeAmount { .minutes(value) }
 }
 
 /// A second of time.
 public struct SecondUnit: TimeUnit {
-    // MARK: TimeUnit
-    
     public static func time(_ value: Int64) -> TimeAmount { .seconds(value) }
 }
 
@@ -65,27 +57,27 @@ public struct SecondUnit: TimeUnit {
 public struct FrequencyTyped<T: TimeUnit>: Frequency {
     /// The frequency at which this work should be repeated.
     let value: Int
-    
+
     /// The day this frequency should start on.
     fileprivate let startDay: Weekday?
-    
+
     /// The hour this frequency should start on.
     fileprivate let startHr: Int?
-    
+
     /// The minute this frequency should start on.
     fileprivate let startMin: Int?
-    
+
     /// The second this frequency should start on.
     fileprivate let startSec: Int?
-    
+
     /// Create a frequency with the given components.
     ///
     /// - Parameters:
-    ///   - value: the value of the unit of time between frequencies.
-    ///   - day: the start day of this frequency.
-    ///   - hr: the start hour of this frequency.
-    ///   - min: the start minute of this frequency.
-    ///   - sec: the start second of this frequency.
+    ///   - value: The value of the unit of time between frequencies.
+    ///   - day: The start day of this frequency.
+    ///   - hr: The start hour of this frequency.
+    ///   - min: The start minute of this frequency.
+    ///   - sec: The start second of this frequency.
     fileprivate init(
         value: Int,
         day: Weekday? = nil,
@@ -99,25 +91,24 @@ public struct FrequencyTyped<T: TimeUnit>: Frequency {
         self.startMin = min
         self.startSec = sec
     }
-    
+
     // MARK: Frequency
-    
+
     public var rate: TimeAmount {
         T.time(Int64(self.value))
     }
-    
+
     public func timeUntilNext() -> TimeAmount {
         let now = Date()
-        let calendar = Calendar.current
         let components = DateComponents(
-            calendar: calendar,
             hour: self.startHr,
             minute: self.startMin,
             second: self.startSec,
             weekday: self.startDay?.rawValue
         )
-        let nextTime = calendar
-            .nextDate(after: now, matching: components, matchingPolicy: .nextTime)!
+
+        let nextTime = Calendar.current
+            .nextDate(after: now, matching: components, matchingPolicy: .nextTime) ?? Date()
         let interval = Int64(nextTime.timeIntervalSince(now) * 1000)
         return .milliseconds(interval)
     }
@@ -131,8 +122,9 @@ public typealias Minutes = FrequencyTyped<MinuteUnit>
 extension Minutes {
     /// When this frequency should first take place.
     ///
-    /// - Parameter sec: a second of a minute (0-59).
-    /// - Returns: a minutely frequency that first takes place at the given component.
+    /// - Parameter sec: A second of a minute (0-59).
+    /// - Returns: A minutely frequency that first takes place at the
+    ///   given component.
     public func at(sec: Int? = nil) -> Minutes {
         Minutes(value: self.value, sec: sec)
     }
@@ -144,9 +136,10 @@ extension Hours {
     /// When this frequency should first take place.
     ///
     /// - Parameters:
-    ///   - min: a minute of an hour (0-59).
-    ///   - sec: a second of a minute (0-59).
-    /// - Returns: an hourly frequency that first takes place at the given components.
+    ///   - min: A minute of an hour (0-59).
+    ///   - sec: A second of a minute (0-59).
+    /// - Returns: An hourly frequency that first takes place at the
+    ///   given components.
     public func at(min: Int? = nil, sec: Int? = nil) -> Hours {
         Hours(value: self.value, min: min, sec: sec)
     }
@@ -158,10 +151,11 @@ extension Days {
     /// When this frequency should first take place.
     ///
     /// - Parameters:
-    ///   - hr: an hour of the day (0-23).
-    ///   - min: a minute of an hour (0-59).
-    ///   - sec: a second of a minute (0-59).
-    /// - Returns: a daily frequency that first takes place at the given components.
+    ///   - hr: An hour of the day (0-23).
+    ///   - min: A minute of an hour (0-59).
+    ///   - sec: A second of a minute (0-59).
+    /// - Returns: A daily frequency that first takes place at the
+    ///   given components.
     public func at(hr: Int? = nil, min: Int? = nil, sec: Int? = nil) -> Days {
         Days(value: self.value, min: min, sec: sec)
     }
@@ -173,11 +167,12 @@ extension Weeks {
     /// When this frequency should first take place.
     ///
     /// - Parameters:
-    ///   - day: a day of the week.
-    ///   - hr: an hour of the day (0-23).
-    ///   - min: a minute of an hour (0-59).
-    ///   - sec: a second of a minute (0-59).
-    /// - Returns: a weekly frequency that first takes place at the given components.
+    ///   - day: A day of the week.
+    ///   - hr: An hour of the day (0-23).
+    ///   - min: A minute of an hour (0-59).
+    ///   - sec: A second of a minute (0-59).
+    /// - Returns: A weekly frequency that first takes place at the
+    ///   given components.
     public func at(_ day: Weekday? = nil, hr: Int? = nil, min: Int? = nil, sec: Int? = nil) -> Weeks {
         Weeks(value: self.value, day: day, hr: hr, min: min, sec: sec)
     }
@@ -193,16 +188,16 @@ public enum Weekday: Int {
 extension Int {
     /// A frequence of weeks.
     public var weeks: Weeks { Weeks(value: self, hr: 0, min: 0, sec: 0) }
-    
+
     /// A frequence of days.
     public var days: Days { Days(value: self, hr: 0, min: 0, sec: 0) }
-    
+
     /// A frequence of hours.
     public var hours: Hours { Hours(value: self, min: 0, sec: 0) }
-    
+
     /// A frequence of minutes.
     public var minutes: Minutes { Minutes(value: self, sec: 0) }
-    
+
     /// A frequence of seconds.
     public var seconds: Seconds { Seconds(value: self) }
 }
