+++ Sources/Alchemy/Jobs/Job/JSONData.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public struct JSONData: Codable {
+    let data: Data
+}
+
+extension JSONData: Parameter {
+    public var value: DatabaseValue { .json(self.data) }
+}
