+++ Sources/Alchemy/Jobs/Job/ScheduledJob.swift
@@ -1,10 +1,3 @@
-//
-//  File.swift
-//  
-//
-//  Created by Chris Anderson on 6/7/20.
-//
-
 import Foundation
 
 public protocol ScheduledJob: Job {
