+++ Sources/Alchemy/Jobs/Queue/Queue.swift
@@ -1,10 +1,3 @@
-//
-//  File.swift
-//  
-//
-//  Created by Chris Anderson on 6/7/20.
-//
-
 import Foundation
 import NIO
