+++ Sources/Alchemy/Jobs/Scheduler.swift
@@ -35,9 +35,9 @@ public struct Scheduler {
             ) { repeatedTask in
                 Log.info("[Scheduler] starting Job `\(name(of: J.self))`.")
                 // For now, never cancel the task.
-                _ = Services.eventLoopGroup.next()
-                    .flatSubmit(job.run)
-                    .map { Log.info("[Scheduler] finished Job `\(name(of: J.self))`.") }
+//                _ = Services.eventLoopGroup.next()
+//                    .flatSubmit(job.run)
+//                    .map { Log.info("[Scheduler] finished Job `\(name(of: J.self))`.") }
             }
         
         return self
